+++ python/acl_anthology/people/person.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 import attrs
 from attrs import define, field, setters
-from enum import Enum
+from enum import StrEnum
 from typing import Any, Iterator, Optional, Sequence, TYPE_CHECKING
 import sys
 
@@ -27,7 +27,7 @@ else:
 
 from ..constants import RE_ORCID, NO_PERSON_ID
 from ..exceptions import AnthologyException, AnthologyInvalidIDError
-from ..utils.attrs import auto_validate_types
+from ..utils.attrs import attach_custom_repr, auto_validate_types, repr_item_ids
 from ..utils.ids import (
     AnthologyID,
     AnthologyIDTuple,
@@ -36,15 +36,15 @@ from ..utils.ids import (
     is_verified_person_id,
     parse_id,
 )
-from . import Name
+from . import Name, ConvertableIntoName
 
 if TYPE_CHECKING:
-    from . import NameSpecification
+    from . import NameSpecification, PersonIndex
     from ..anthology import Anthology
     from ..collections import Paper, Volume
 
 
-class NameLink(Enum):
+class NameLink(StrEnum):
     """How a Name was connected to a Person."""
 
     EXPLICIT = "explicit"
@@ -52,6 +52,9 @@ class NameLink(Enum):
 
     INFERRED = "inferred"
     """Name was connected to this Person via slug matching heuristic."""
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}.{self.name}"
 
 
 def _name_list_converter(
@@ -82,7 +85,7 @@ def _update_person_index(person: Person, attr: attrs.Attribute[Any], value: str)
 
     Intended to be called from `on_setattr` of an [attrs.field][].
     """
-    index = person.parent.people
+    index = person.parent
     if attr.name == "id":
         index._update_id(person.id, value)
     elif attr.name == "orcid":
@@ -90,6 +93,7 @@ def _update_person_index(person: Person, attr: attrs.Attribute[Any], value: str)
     return value
 
 
+@attach_custom_repr
 @define(field_transformer=auto_validate_types)
 class Person:
     """A natural person.
@@ -102,7 +106,7 @@ class Person:
     Attributes:
         id: A unique ID for this person.  Do not change this attribute directly; use [`change_id()`][acl_anthology.people.person.Person.change_id], [`make_explicit()`][acl_anthology.people.person.Person.make_explicit], or [`merge_into()`][acl_anthology.people.person.Person.merge_into] instead.
         parent: The parent Anthology instance to which this person belongs.
-        item_ids: A list of volume and/or paper IDs this person has authored or edited.
+        item_ids: An unordered set of volume and/or paper IDs this person has authored or edited.
         orcid: The person's ORCID.
         comment: A comment for disambiguation purposes.
         degree: The person's institution of highest degree, for disambiguation purposes.
@@ -110,13 +114,18 @@ class Person:
         is_explicit: If True, this person's ID is explicitly defined in `people.yaml`.  You probably want to use [`make_explicit()`][acl_anthology.people.person.Person.make_explicit] rather than change this attribute.
     """
 
-    id: str = field(on_setattr=[setters.validate, _update_person_index])
-    parent: Anthology = field(repr=False, eq=False)
-    _names: list[tuple[Name, NameLink]] = field(
-        factory=list, converter=_name_list_converter
+    id: str = field(
+        on_setattr=[setters.validate, _update_person_index],
+        metadata={"repr_omits_field_name": True},
     )
-    item_ids: list[AnthologyIDTuple] = field(
-        factory=list, repr=lambda x: f"<list of {len(x)} AnthologyIDTuple objects>"
+    parent: PersonIndex = field(repr=False, eq=False)
+    _names: list[tuple[Name, NameLink]] = field(
+        factory=list,
+        converter=_name_list_converter,
+        metadata={"repr_omits_field_name": True},
+    )
+    item_ids: set[AnthologyIDTuple] = field(
+        factory=set, converter=set, repr=repr_item_ids
     )
     orcid: Optional[str] = field(
         default=None,
@@ -126,7 +135,7 @@ class Person:
     degree: Optional[str] = field(default=None)
     similar_ids: list[str] = field(factory=list)
     _disable_name_matching: Optional[bool] = field(default=False, converter=bool)
-    is_explicit: Optional[bool] = field(default=False, converter=bool)
+    is_explicit: Optional[bool] = field(default=False, converter=bool, repr=False)
 
     def __eq__(self, other: object) -> bool:
         if not isinstance(other, Person):
@@ -137,18 +146,26 @@ class Person:
         return hash(self.id)
 
     @property
+    def root(self) -> Anthology:
+        """The Anthology instance to which this object belongs."""
+        return self.parent.parent
+
+    @property
     def names(self) -> list[Name]:
         """A list of all names associated with this person."""
         return [name for (name, _) in self._names]
 
     @names.setter
-    def names(self, values: list[Name]) -> None:
+    def names(self, values: list[ConvertableIntoName]) -> None:
         for name, _ in self._names:
-            self.parent.people._remove_name(self.id, name)
-        for name in values:
-            self.parent.people._add_name(self.id, name)
-        self._names = _name_list_converter(values)
-        self.parent.people._reresolve_items_for_person(self)
+            self.parent._remove_name(self.id, name)
+        names = []
+        for ci_name in values:
+            name = Name.from_(ci_name)
+            self.parent._add_name(self.id, name)
+            names.append((name, NameLink.EXPLICIT))
+        self._names = names
+        self.parent._reresolve_items_for_person(self)
 
     @property
     def disable_name_matching(self) -> bool:
@@ -161,7 +178,7 @@ class Person:
     @disable_name_matching.setter
     def disable_name_matching(self, value: Optional[bool]) -> None:
         self._disable_name_matching = value
-        self.parent.people._reresolve_items_for_names(self.names)
+        self.parent._reresolve_items_for_names(self.names)
 
     @property
     def canonical_name(self) -> Name:
@@ -174,8 +191,8 @@ class Person:
             raise ValueError(f"No names defined for person '{self.id}'")
 
     @canonical_name.setter
-    def canonical_name(self, name: Name) -> None:
-        self._set_canonical_name(name)
+    def canonical_name(self, name: ConvertableIntoName) -> None:
+        self._set_canonical_name(Name.from_(name))
 
     def _set_canonical_name(self, name: Name, inferred: bool = False) -> None:
         """Set the canonical name for this person.
@@ -189,10 +206,11 @@ class Person:
         link_type = NameLink.INFERRED if inferred else NameLink.EXPLICIT
         if not self.has_name(name):
             self._names = [(name, link_type)] + self._names
+            self.parent._add_name(self.id, name)
         else:
             self._names = [(name, link_type)] + [x for x in self._names if x[0] != name]
 
-    def add_name(self, name: Name, inferred: bool = False) -> None:
+    def add_name(self, name: ConvertableIntoName, inferred: bool = False) -> None:
         """Add a name for this person.
 
         Parameters:
@@ -200,18 +218,19 @@ class Person:
             inferred: If True, will be marked as `NameLinkingType.INFERRED`, which will e.g. cause this name to not be written to `people.yaml`.  Used when building the [`PersonIndex`][acl_anthology.people.index.PersonIndex] from the XML data; you probably don't want to set this manually.  Defaults to False.
         """
         link_type = NameLink.INFERRED if inferred else NameLink.EXPLICIT
+        name = Name.from_(name)
         if not self.has_name(name):
             self._names.append((name, link_type))
-            self.parent.people._add_name(self.id, name)
+            self.parent._add_name(self.id, name)
             if not inferred:
-                self.parent.people._reresolve_items_for_names([name])
+                self.parent._reresolve_items_for_names([name])
         elif (name, link_type) not in self._names:
             # ensure that name is re-inserted at same position
             idx = self.names.index(name)
             del self._names[idx]
             self._names.insert(idx, (name, link_type))
 
-    def remove_name(self, name: Name) -> None:
+    def remove_name(self, name: ConvertableIntoName) -> None:
         """Remove an explicit name for this person.
 
         Warning:
@@ -223,12 +242,13 @@ class Person:
         Raises:
             ValueError: If this name was not explicitly linked to this person.
         """
+        name = Name.from_(name)
         self._names.remove((name, NameLink.EXPLICIT))
-        self.parent.people._remove_name(self.id, name)
-        self.parent.people._reresolve_items_for_person(self)
-        self.parent.people._reresolve_items_for_names([name])
+        self.parent._remove_name(self.id, name)
+        self.parent._reresolve_items_for_person(self)
+        self.parent._reresolve_items_for_names([name])
 
-    def has_name(self, name: Name) -> bool:
+    def has_name(self, name: ConvertableIntoName) -> bool:
         """
         Parameters:
             name: Name to be checked.
@@ -236,7 +256,9 @@ class Person:
         Returns:
             True if the given name can refer to this person.
         """
-        return any(existing_name == name for (existing_name, _) in self._names)
+        return any(
+            existing_name == Name.from_(name) for (existing_name, _) in self._names
+        )
 
     def change_id(self, new_id: str) -> None:
         """Change this person's ID.
@@ -250,7 +272,7 @@ class Person:
             AnthologyException: If `self.explicit` is False.
             AnthologyInvalidIDError: If the supplied ID is not well-formed, or if it already exists in the PersonIndex.
         """
-        if new_id in self.parent.people:
+        if new_id in self.parent:
             exc = AnthologyInvalidIDError(new_id, f"Person ID already exists: {new_id}")
             exc.add_note("Did you want to use merge_into() instead?")
             raise exc
@@ -288,12 +310,12 @@ class Person:
         if self.is_explicit:
             raise AnthologyException(f"Person '{self.id}' is already explicit")
         if new_id is None:
-            new_id = self.parent.people.generate_person_id(self)
+            new_id = self.parent.generate_person_id(self)
         elif not is_verified_person_id(new_id):
             raise AnthologyInvalidIDError(
                 new_id, f"Not a valid verified-person ID: {new_id}"
             )
-        elif new_id in self.parent.people:
+        elif new_id in self.parent:
             raise AnthologyException(f"ID already exists in the index: {new_id}")
 
         namespecs = list(self.namespecs())
@@ -379,7 +401,7 @@ class Person:
         """Returns an iterator over all Anthology items associated with this person, regardless of their type."""
         # TODO: This does not consider talks yet!
         for anthology_id in self.item_ids:
-            item = self.parent.get(anthology_id)
+            item = self.root.get(anthology_id)
             if item is None:
                 raise ValueError(
                     f"Person {self.id} lists associated item {build_id_from_tuple(anthology_id)}, which doesn't exist"
@@ -401,7 +423,7 @@ class Person:
         for anthology_id in self.item_ids:
             paper_id = anthology_id[-1]
             if paper_id is not None:
-                paper = self.parent.get_paper(anthology_id)
+                paper = self.root.get_paper(anthology_id)
                 if paper is None:
                     raise ValueError(
                         f"Person {self.id} lists associated paper {build_id_from_tuple(anthology_id)}, which doesn't exist"
@@ -413,7 +435,7 @@ class Person:
         for anthology_id in self.item_ids:
             paper_id = anthology_id[-1]
             if paper_id is None:
-                volume = self.parent.get_volume(anthology_id)
+                volume = self.root.get_volume(anthology_id)
                 if volume is None:
                     raise ValueError(
                         f"Person {self.id} lists associated volume {build_id_from_tuple(anthology_id)}, which doesn't exist"